// src/api/openapi/models/articleUpdateFull.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ArticleUpdateFull { 
    category_ids: Array<number>;
    description_de: string;
    description_it: string;
    favorite: boolean | null;
    mod_number: string;
    name_de: string;
    name_it: string;
    price: number;
    unit_id: number;
    vat_id: number;
}